Update code and textures to Minecraft 1.15.2

+++ src/main/java/chylex/hee/client/render/territory/components/SkyCubeStatic.kt
@@ -6,6 +6,6 @@ class SkyCubeStatic(
 	override val texture: ResourceLocation = DEFAULT_TEXTURE,
 	override val color: Vec3d = DEFAULT_COLOR,
 	override val alpha: Float = DEFAULT_ALPHA,
-	override val rescale: Double = DEFAULT_RESCALE,
-	override val distance: Double = DEFAULT_DISTANCE
+	override val rescale: Float = DEFAULT_RESCALE,
+	override val distance: Float = DEFAULT_DISTANCE
 ) : SkyCubeBase()
